requirements.txt
aioredis==1.3.0
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.1
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
gunicorn==20.0.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
whitenoise==5.0.1
zope.interface==4.6.0
